requirements.txt
Django>=5.2.7
djangorestframework>=3.14.0
asgiref>=3.10.0
sqlparse>=0.5.3
typing_extensions>=4.15.0
